:art: Organize exercises per section
The Udemy course is split in multiple sections per major topic. All the
exercises will be coded in this same repository, however each section
will have a separate directory named after the course section.

This structure has been chosen instead of separate repositories to
group all exercises due to the simplicity of the exercises and due to
the fact that most exercises - if not all of them - are built around web
pages instead of web applications such as Angular or React apps.